improve v4l2 behavior for shared buffers
when the decoder is using pad_alloc(), v4l2sink would behave badly if the number of buffers ('queue-size' property) was not high enough to account for all the buffers needed by the decoder, and other elements (such as queues) between the decoder and v4l2sink.  This patch slightly increases the default number of buffers, and changes v4l2sink to drop frames rather than return an error in case the number of buffers is not high enough.